Game Manager Department Arrangement SRP Principle
